Drop v1.0 and v1.1
