+++ app/controllers/api/v1x2/mixins/index_mixin.rb
@@ -1,5 +1,5 @@
 module Api
-  module V1x0
+  module V1x2
     module Mixins
       module IndexMixin
         def scoped(relation)
