+++ app/controllers/api/v1x2/stageaction/_footer.html.erb
