+++ app/controllers/api/v1x2/stageaction/result.html.erb
