+++ app/controllers/api/v1x2/stageaction/show.html.erb
