+++ app/controllers/api/v1x2/templates_controller.rb
@@ -1,5 +1,5 @@
 module Api
-  module V1x0
+  module V1x2
     class TemplatesController < ApplicationController
       include Mixins::IndexMixin
